Implement tools to produce network graphs in memory net examples
This PR adds some utility functions to take snapshots of the network as a graph at any point during a memory net run. The utility functions provide the ability to take snapshots in separate named sequences if you want to graph different networks or elements of the networks separately. The graph files are stored in the temporary directory. Finally a Python script is included that will read the graphs for each named sequence and render them using a Force-Directed graph layout algorithm to the provided (or default) output folder.

Supported:
 - Network graphs with all connections and/or neighbour connections
- Message propagation onto of a network graph