submit_block returns OK if block was submitted for tari or monero
Pool software expects for the submit block call to succeed if a block it
just mined was able to be submitted to the blockchain. Previously, if
the block was successfully submitted to the tari blockchain but not the
monero blockchain, the call would fail.

This PR changes the behaviour of submit_block to succeed if either a
tari or monero block succeeded.