chore: add versions to all hash domains (#5971)
Description
---
Adds versioning to all hash domains.

Closes #5970.

Motivation and Context
---
Some hash function definitions rely on a default version number that is
added during macro invocation. This isn't inherently a problem, but
reduces clarity and could result in breaking changes if this default is
ever changed. It seems prudent to add a specific version in all cases.

How Has This Been Tested?
---
Existing tests pass.

What process can a PR reviewer use to test or verify this change?
---
Confirm that the added version numbers correspond to the
[default](https://github.com/tari-project/tari-crypto/blob/91d1e779e565dc816a54ab78368d44efcae78a23/src/hashing.rs#L616).